add some comments. fix for phi-3.5-moe custom router

+++ vllm/model_executor/layers/fused_moe/router/custom_routing_router.py
@@ -50,4 +50,4 @@ class CustomRoutingRouter(BaseRouter):
             renormalize=self.renormalize,
         )
 
-        return topk_weights, topk_ids
+        return topk_weights.to(torch.float32), topk_ids
